requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bleach==6.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.10.0
packaging==24.2
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stripe==7.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.2.1
